extract MealPlan component from PlanMeals, create utils folder, put plans.ts with meal plans data there
