+++ src/utils/plans.ts
@@ -1,0 +1,56 @@
+export const PLANS = [
+    {
+      id: 476,
+      name: 'Keto Meal Plan',
+      description:
+        'a dietary plan that focuses on consuming high-fat, low-carbohydrate foods to induce a state of ketosis in the body for weight loss and increased energy',
+    },
+    {
+      id: 128,
+      name: '1500 Calorie Meal Plan',
+      description:
+        'a dietary plan designed to provide a daily intake of 1500 calories, typically aiming for balanced nutrition while creating a calorie deficit for weight management or certain health goals',
+    },
+    {
+      id: 124,
+      name: 'Vegetarian Meal Plan',
+      description:
+        'a dietary plan that excludes meat and seafood, focusing on plant-based foods such as fruits, vegetables, grains, legumes, and dairy products for nutrition and sustainability reasons',
+    },
+    {
+      id: 409,
+      name: 'week 1 Vegan',
+      description:
+        'is a meal plan designed for individuals following a vegan lifestyle, incorporating plant-based foods and excluding all animal products, including meat, dairy, eggs, and honey',
+    },
+    {
+      id: 578,
+      name: 'Whole30 Meal Plan',
+      description:
+        "a 30-day dietary plan that eliminates certain food groups such as grains, dairy, legumes, sugar, and processed foods, aiming to reset the body's relationship with food and improve overall health",
+    },
+    {
+      id: 161,
+      name: 'Almost Meatless',
+      description:
+        'a meal plan focuses on reducing meat consumption by incorporating a variety of plant-based foods while occasionally including small portions of meat or fish for a balanced and sustainable approach to eating',
+    },
+    {
+      id: 1272,
+      name: 'Super budget week',
+      description:
+        'a meal plan is designed to provide affordable and cost-effective meals, utilizing inexpensive ingredients and budget-friendly recipes without compromising on taste and nutrition',
+    },
+    {
+      id: 217,
+      name: 'smoothie and salad week',
+      description:
+        'meal plan that emphasizes the consumption of nutrient-rich smoothies and salads, providing a refreshing and health-conscious approach to eating with a focus on fruits, vegetables, and greens',
+    },
+    {
+      id: 1745,
+      name: 'easy-carb-week',
+      description:
+        'a meal plan is designed to provide simple and convenient meals with a focus on carbohydrate-rich foods, offering quick and satisfying options for individuals seeking an easy approach to their carbohydrate intake',
+    },
+  ];
